requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
faiss-cpu==1.7.4
sentence-transformers==2.3.1
beautifulsoup4==4.12.3
requests==2.31.0
streamlit==1.30.0
lxml==5.1.0
pydantic==2.5.3
python-multipart==0.0.6
selenium==4.16.0
